+++ workbench/app/Models/User.php
@@ -10,7 +10,7 @@ use Illuminate\Foundation\Auth\User as Authenticatable;
 use Illuminate\Notifications\Notifiable;
 use Workbench\Database\Factories\UserFactory;
 
-#[Fillable(['name', 'email', 'password'])]
+#[Fillable(['name', 'email', 'password', 'country_code', 'phone'])]
 #[Hidden(['password', 'remember_token'])]
 class User extends Authenticatable
 {