+++ workbench/routes/web.php
@@ -2,6 +2,7 @@
 
 use Illuminate\Support\Facades\Route;
 
-Route::get('/', function () {
-    return view('welcome');
-});
+Route::redirect('/', '/users');
+
+Route::livewire('/users', 'pages::users')->name('users.index');
+Route::livewire('/phones', 'pages::phone-checker')->name('phones.checker');